* src/manage_sql.c (manage_set_config_preference): Handle osp config preference case.
git-svn-id: svn+ssh://scm.wald.intevation.org/openvas/trunk/openvas-manager@20353 423fd1db-d629-0410-8442-d21db03e70f4

+++ src/manage_sql.c
@@ -28588,15 +28588,23 @@ manage_set_config_preference (config_t config, const char* nvt, const char* name
   quoted_value = sql_quote ((gchar*) value);
   g_free (value);
 
-  sql ("DELETE FROM config_preferences"
-       " WHERE config = %llu AND type %s AND name = '%s'",
-       config,
-       nvt ? "= 'PLUGINS_PREFS'" : "= 'SERVER_PREFS'",
-       quoted_name);
-  sql ("INSERT INTO config_preferences"
-       " (config, type, name, value) VALUES (%llu, %s, '%s', '%s');",
-       config, nvt ? "'PLUGINS_PREFS'" : "'SERVER_PREFS'", quoted_name,
-       quoted_value);
+  if (config_type (config) > 0)
+    sql ("UPDATE config_preferences SET value = '%s'"
+         " WHERE config = %llu AND name = '%s';",
+         quoted_value, config, quoted_name);
+  else
+    {
+      /* nvt prefs are not present on first modification. */
+      sql ("DELETE FROM config_preferences"
+           " WHERE config = %llu AND type %s AND name = '%s'",
+           config,
+           nvt ? "= 'PLUGINS_PREFS'" : "= 'SERVER_PREFS'",
+           quoted_name);
+      sql ("INSERT INTO config_preferences"
+           " (config, type, name, value) VALUES (%llu, %s, '%s', '%s');",
+           config, nvt ? "'PLUGINS_PREFS'" : "'SERVER_PREFS'", quoted_name,
+           quoted_value);
+    }
 
   sql ("COMMIT;");
